+++ src/components/commons/card/StudyAcceptCard.tsx
@@ -1,7 +1,7 @@
 import Image from "next/image";
-import StudyApplyInfo from "./StudyApplyInfo";
-import StudyCardButton from "./StudyCardButton";
+import StudyApplyInfo from "./element/StudyApplyInfo";
 import { business as sampleImage } from "../../../../public/icons/icons";
+import StudyCardButton from "./element/StudyCardButton";
 
 export default function StudyAcceptCard({ status }: { status: "APPLIED" | "APPLIED_VIEW" | "ACCEPTED" }) {
   return (
