+++ src/components/commons/card/element/StudyApplyInfo.tsx
