+++ src/components/commons/tag/TagLight.tsx
@@ -1,0 +1,5 @@
+import TagBase, { ITagProps } from "../TagBase";
+
+export function TagLight({ children }: ITagProps) {
+  return <TagBase className="border-main-500 bg-main-100 text-main-500">{children}</TagBase>;
+}
