+++ src/components/commons/tag/TagMain.tsx
@@ -1,0 +1,5 @@
+import TagBase, { ITagProps } from "../TagBase";
+
+export function TagMain({ children }: ITagProps) {
+  return <TagBase className="border-main-500 bg-main-500 text-white">{children}</TagBase>;
+}
